Seperate functional tests out by command
This just involves taking all the code in `tests/integration/test___main__.py`
and splitting into separate files in `tests/functional`.
