+++ tests/functional/test_test_connection.py
@@ -1,0 +1,12 @@
+from ehrql.__main__ import main
+
+
+def test_test_connection(mssql_database, capsys):
+    env = {
+        "BACKEND": "ehrql.backends.tpp.TPPBackend",
+        "DATABASE_URL": mssql_database.host_url(),
+    }
+    argv = ["test-connection"]
+    main(argv, env)
+    out, _ = capsys.readouterr()
+    assert "SUCCESS" in out
